MASSIVE speedup of grammar/dictionary initialization (20+ second improvement)
On my machine, this used to take 27 seconds. Now it takes less than 5 seconds.

The slowdown was mainly caused by invoking the garbage collector manually in a loop. Premature optimization is the root of all evil.

Also, deserializing the grammar file turns out to be slower than reading the text file manually (with the current dictionary structure anyway).